requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
aws-cfn-bootstrap==2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cssselect==1.2.0
docutils==0.14
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
Flask==2.2.1
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lockfile==0.11.0
lxml==4.9.2
MarkupSafe==2.1.1
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
openai==0.26.1
Pillow==9.4.0
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
simplejson==3.2.0
six==1.16.0
soupsieve==2.3.2.post1
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0